requirements.txt
pandas==1.2.2
mplfinance==0.12.7a17
yfinance==0.1.55
plotly==4.14.3
PyYAML==5.4.1
rich==10.3.0
